// src/models/bg.props.ts
export default interface StylesProps {
  backgroundRepeat: string;
  backgroundImage: string;
  height?: string;
  resize: string;
  minHeight: string;
}

// TODO - instert all properties CSS, because is generic
export interface IStylesDefault {
  background: string;
  color: string;
}